refactor: CarRacingManager 객체를 생성하여 의존성 주입 및 주요 로직 실행을 담당하도록 수정
프로그램의 사이즈가 더 커지게 되면 main 메서드의 길이가 너무 길어질 것을 우려하여, Manager 객체를 정의하여 애플리케이션 실행의 전반적인 로직을 담당하도록 변경하고, main함수에선 간단하게 CarRacingManager 객체의 execute 메서드를 호출함으로서 애플리케이션의 의존성 주입과 주요 로직 실행을 담당하도록 개선
